requirements.txt
streamlit
newspaper3k 
textblob
requests